+++ .settings/stm32cubeide.project.prefs
@@ -1,4 +1,4 @@
-66BE74F758C12D739921AEA421D593D3=0
+66BE74F758C12D739921AEA421D593D3=1
 8DF89ED150041C4CBC7CB9A9CAA90856=B5019EF660F2E1CF83D068680A386D93
 DC22A860405A8BF2F2C095E5B6529F12=B5019EF660F2E1CF83D068680A386D93
 eclipse.preferences.version=1
