+++ Debug/CPP/SRC/main.d
@@ -1,4 +1,4 @@
-CPP/main.o: ../CPP/main.cpp ../Core/Inc/main.h \
+CPP/SRC/main.o: ../CPP/SRC/main.cpp ../Core/Inc/main.h \
  ../Drivers/STM32G0xx_HAL_Driver/Inc/stm32g0xx_hal.h \
  ../Core/Inc/stm32g0xx_hal_conf.h \
  ../Drivers/STM32G0xx_HAL_Driver/Inc/stm32g0xx_hal_rcc.h \
