+++ Debug/CPP/SRC/main.su
@@ -1,0 +1,3 @@
+../CPP/SRC/main.cpp:13:6:void maincpp()	16	static
+../CPP/SRC/main.cpp:40:6:void HAL_GPIO_EXTI_Rising_Callback(uint16_t)	16	static
+../CPP/SRC/main.cpp:49:6:void HAL_GPIO_EXTI_Falling_Callback(uint16_t)	16	static
