+++ Debug/objects.list
@@ -1,4 +1,4 @@
-"./CPP/main.o"
+"./CPP/SRC/main.o"
 "./Core/Src/main.o"
 "./Core/Src/stm32g0xx_hal_msp.o"
 "./Core/Src/stm32g0xx_it.o"
